agents/content_generation/requirements.txt
pika==1.3.2
requests==2.31.0
python-dotenv==1.0.0
openai==0.27.8
prometheus-client==0.17.1